// src/types/ProjectsItem.ts
export interface ProjectsItem {
  img: string;
  url: string;
  title: string;
  description: string;
  id: number;
  techs: string[];
  delay: number;
}

// const projectsItem: ProjectsItem[] = [
//   {
//     img: "/img/img-projects/project-2.png",
//     title: "Mini Game Tic Tac Toe",
//     url: "https://alvinnes.github.io/tic-tac-toe.github.io/",
//     description:
//       "A classic Tic Tac Toe game built with JavaScript, allowing two players to compete in a simple and fun interface.",
//     id: 1,
//     techs: ["HTML", "Tailwind CSS", "React"],
//     delay: 0.1,
//   },
//   {
//     img: "/img/img-projects/project-3.png",
//     title: "Simple Kalkulator App",
//     url: "/projects/kalkulator/index.html",
//     description:
//       "A basic calculator web app that performs standard arithmetic operations with a user-friendly UI.",
//     id: 2,
//     techs: ["HTML", "CSS", "Javascript"],
//     delay: 0.2,
//   },
//   {
//     img: "/img/img-projects/project-4.png",
//     title: "Simple To-do-list App",
//     url: "https://to-do-list-kappa-two-32.vercel.app/",
//     description:
//       "A simple and intuitive to-do list app for organizing tasks efficiently, built with clean design and local storage support.",
//     id: 3,
//     techs: ["HTML", "CSS", "React"],
//     delay: 0.3,
//   },
//   {
//     img: "/img/img-projects/project-5.png",
//     title: "Wibuku",
//     url: "https://wibuku-alvinnes-projects.vercel.app/",
//     description:
//       "An online digital book catalog where users can explore and manage their reading list.",
//     id: 4,
//     techs: ["HTML", "Tailwindcss", "React", "Typescript"],
//     delay: 0.1,
//   },
//   {
//     img: "/img/img-projects/simple-landing-page.png",
//     title: "Simple Landing Page",
//     url: "https://alvinnes.github.io/profile.github.io/",
//     description:
//       "A personal landing page showcasing profile information and web development portfolio.",
//     id: 5,
//     techs: ["HTML", "CSS"],
//     delay: 0.2,
//   },
//   {
//     img: "/img/img-projects/caffinity.png",
//     title: "Coffe Shop Caffinity",
//     url: "https://caffinity-six.vercel.app/",
//     description:
//       "A modern coffee shop website with an interactive menu, product filtering, and elegant UI design.",
//     id: 6,
//     techs: ["HTML", "CSS", "React", "Typescript"],
//     delay: 0.3,
//   },
// ];

// export default projectsItem;